Update Hero Skill

+++ Assets/BaseGame/Scripts/Core/StatusEffect/IceSlowStatusEffect.cs
@@ -9,7 +9,7 @@ namespace Core.GameStatusEffect
         [field: SerializeField] public Color SlowColor { get; private set; } = new Color(0f, 0.5f, 1f);
         [field: SerializeField] public Color NormalColor { get; private set; } = Color.white;
 
-        public IceSlowStatusEffect( float slowAmount, float duration) : base(slowAmount, duration)
+        public IceSlowStatusEffect( int slowAmount, float duration) : base(slowAmount, duration)
         {
             
         }
